requirements.txt
# Universal Project Launcher - Dependencies
# Python 3.8+ required

# Core dependencies
PyYAML>=6.0
python-dotenv>=1.0.0

# GUI and system tray (optional, for full launcher)
pystray>=0.19.0
Pillow>=10.0.0

# Terminal management (optional, for advanced features)
libtmux>=0.20.0

# CLI enhancements
typer>=0.9.0
rich>=13.0.0

# HTTP client (for health checks)
httpx>=0.24.0